+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aguspe/tiler-server
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @aguspe/tiler-viewer@1.0.3
+  - @aguspe/tiler-editor@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes